requirements.txt
fastapi
python-dotenv
motor
pyjwt
uvicorn
jinja2
pydantic
pydantic[email]
httpx